--- routes/about.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-router.get('/', (req, res) => {
-	return res.renderFile('about.njk');
-});
-
-module.exports = router;
