--- routes/login.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-router.get('/', (req, res) => {
-	if (req.loggedIn) return res.redirect('/');
-	res.renderFile('login.njk');
-});
-
-module.exports = router;
