--- routes/submissions.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-router.get('/', (req, res) => {
-	return res.renderFile('submissions.njk');
-});
-
-module.exports = router;
