--- routes/success.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-router.get('/', (req, res) => {
-	// TODO: Rename this to /quiz/success
-	return res.renderFile('events/quiz_success.njk');
-});
-
-module.exports = router;
